requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
pymysql
Werkzeug
requests
PyJWT
reportlab
Pillow
pyinstaller
